factored out mute button into own component and functional
